requirements.txt
imgaug==0.2.9
matplotlib==3.1.0
scikit_image==0.15.0
torchvision==0.3.0
tqdm==4.32.1
numpy==1.16.3
opencv_python==4.1.0.25
scipy==1.3.0
pycocotools==2.0.0
scikit_multilearn==0.2.0
torch==1.1.0
colorlog==4.0.2
Pillow==7.0.0
python-dotenv==0.12.0
skimage==0.0
scikit_learn==0.22.2.post1
tensorboardX==2.0
typing==3.7.4.1
